feature: not allowing users to select duplicate majors in diff modals
